refactor(shared-switch): unify platform implementation
